# api/openapi.yaml
openapi: 3.0.3
info:
  title: MIMO OTA Control API
  version: 0.9.0
servers:
  - url: http://localhost:3000
paths:
...
  /api/v1/test-executions:
    get:
      summary: List execution history (ARCH-1 S2, sourced from test_executions)
      description: |
        One row per execution from the test_executions table itself
        (case-runner / commissioning / diagnostics; VRT rows excluded),
        newest first by executed_at. Includes RUNNING rows — the test case
        library uses status=running to restore the active-run badge after
        navigation (Codex #237 C3). Row id is TestExecution.id and can be
        passed directly as a report's test_execution_ids.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: status
          in: query
          description: |
            Row status. Covers every value test_executions.status can hold
            for non-VRT rows: pending (session row not yet run), running,
            completed, failed, cancelled, skipped (ad-hoc phase skipped).
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled, skipped]
        - name: executed_by
          in: query
          description: |
            Filter by execution chain (test_case_runner / test_plan_runner /
            commissioning_api / commissioning_adhoc). Repeatable — pass the
            parameter once per chain to match any of them.

            Narrowing MUST happen here rather than client-side: client-side
            filtering runs after the limit window, so a chain with many rows
            crowds out the wanted ones. Two instances so far — the case
            library's active-run recovery passes test_case_runner (stale
            plan-runner running rows would crowd it out), and the
            pending-report list passes the three formal chains (ad-hoc
            diagnostics would crowd out archivable executions).
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Execution history page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestExecutionListResponse'
...
  /api/v1/test-plans/cases/executions/{execution_id}:
    get:
      summary: Get one TestCase execution status and sanitized SCPI evidence
      parameters:
        - name: execution_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Execution status with P1-47C evidence when available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseExecutionStatus'
        '404':
          description: Execution not found
...
components:
  schemas:
...
    TestExecutionItem:
      type: object
      description: |
        ARCH-1 S2: one execution row from the test_executions table itself
        (case-runner / plan-runner step / commissioning / ad-hoc diagnostic;
        VRT rows excluded via mode IS NULL). Includes running rows. Mirrors
        api-service ExecutionHistoryItem. phases_* null = that execution
        chain does not record phase progress; validation_pass null = not
        judged.
      properties:
        id:
          type: string
        case_name:
          type: string
          nullable: true
        source_test_case_id:
          type: string
          nullable: true
        status:
          type: string
        phases_total:
          type: integer
          nullable: true
        phases_done:
          type: integer
          nullable: true
        phases_failed:
          type: integer
          nullable: true
        duration_sec:
          type: number
          nullable: true
        started_at:
          type: string
          nullable: true
        completed_at:
          type: string
          nullable: true
        executed_by:
          type: string
          nullable: true
        error_message:
          type: string
          nullable: true
        validation_pass:
          type: boolean
          nullable: true
        execution_classification:
          type: string
          enum: [formal, diagnostic, legacy]
        failure_alert_outcome:
          type: string
          nullable: true
          description: >-
            P2-34: execution-failure alert publication outcome recorded on
            this row (published | duplicate | failed). null = not recorded
            (rows predating P2-34 / record write failed / not applicable) —
            null does NOT mean the alert was published.
      required:
        - id
        - case_name
        - source_test_case_id
        - status
        - phases_total
        - phases_done
        - phases_failed
        - duration_sec
        - started_at
        - completed_at
        - executed_by
        - error_message
        - validation_pass
        - execution_classification
        - failure_alert_outcome
    TestExecutionListResponse:
      type: object
      properties:
        total:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/TestExecutionItem'
      required: [total, items]
    ScpiEvidenceEnvironment:
      type: object
      additionalProperties: true
      properties:
        instrument_id:
          type: string
        instrument:
          type: string
        model:
          type: string
          nullable: true
        firmware_version:
          type: string
          nullable: true
        captured_from_live_connection:
          type: boolean
      required: [instrument_id, instrument, captured_from_live_connection]
    ScpiEvidenceRequirement:
      type: object
      properties:
        requirement_id:
          type: string
        evidence_key:
          type: string
        requested: {}
        required_evidence_level:
          type: string
          enum: [E0, E1, E2, E3, E4]
      required: [requirement_id, evidence_key, requested, required_evidence_level]
    ScpiEvidenceItem:
      type: object
      properties:
        requirement_id:
          type: string
        instrument:
          type: string
        evidence_key:
          type: string
        requested: {}
        command_sent:
          type: string
          nullable: true
        readback: {}
        exchange_ids:
          type: array
          items:
            type: string
        evidence_level:
          type: string
          enum: [E0, E1, E2, E3, E4]
        source_reference:
          type: string
          nullable: true
        verdict:
          type: string
          enum: [passed, rejected, unknown]
        reason:
          type: string
      required:
        - requirement_id
        - instrument
        - evidence_key
        - requested
        - command_sent
        - readback
        - exchange_ids
        - evidence_level
        - source_reference
        - verdict
        - reason
    ExecutionScpiEvidence:
      type: object
      description: |
        Sanitized P1-47C summary. Raw replies stay in scpi.log and are linked
        by execution_id + exchange_ids. Any missing/unknown/rejected mandatory
        item makes formal_acceptance false.
      properties:
        schema_version:
          type: integer
        execution_id:
          type: string
          format: uuid
        environments:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ScpiEvidenceEnvironment'
        required:
          type: array
          items:
            $ref: '#/components/schemas/ScpiEvidenceRequirement'
        items:
          type: array
          items:
            $ref: '#/components/schemas/ScpiEvidenceItem'
        missing_requirements:
          type: array
          items:
            type: string
        formal_verdict:
          type: string
          enum: [passed, rejected, unknown]
        formal_acceptance:
          type: boolean
        reason:
          type: string
      required:
        - schema_version
        - execution_id
        - environments
        - required
        - items
        - missing_requirements
        - formal_verdict
        - formal_acceptance
        - reason
    CaseExecutionStatus:
      type: object
      properties:
        execution_id:
          type: string
          format: uuid
        status:
          type: string
        source_test_case_id:
          type: string
          nullable: true
        phase_progress:
          type: array
          items:
            type: object
            additionalProperties: true
        failed_phase:
          type: string
          nullable: true
        error_message:
          type: string
          nullable: true
        started_at:
          type: string
          nullable: true
        completed_at:
          type: string
          nullable: true
        scpi_evidence:
          allOf:
            - $ref: '#/components/schemas/ExecutionScpiEvidence'
          nullable: true
      required: [execution_id, status, phase_progress]
...